# assets/openapi/sonos-hub.v1.yaml
openapi: 3.0.3
info:
  title: Sonos Smart Routines Hub API
  version: '1.0.0'
  description: |
    REST API for the Sonos Smart Routines Hub - a LAN-only home automation system for
    Sonos speakers. Provides device discovery, scene/routine management, music playback
    control, and integration with Apple Music. Designed for local-first operation with
    observability via Prometheus metrics and structured JSON logging.
servers:
  - url: http://hub.local:8787
...
paths:
...
  /v1/scenes:
    get:
      operationId: listScenes
      tags: [scenes]
      summary: List scenes
      description: Get a paginated list of all scenes
      parameters:
        - in: query
          name: limit
          description: Maximum number of scenes to return
          schema: { type: integer }
        - in: query
          name: offset
          description: Number of scenes to skip for pagination
          schema: { type: integer }
      responses:
        '200':
          description: List of scenes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScenesListResponse' }
    post:
      operationId: createScene
      tags: [scenes]
      summary: Create scene
      description: Create a new scene with playback configuration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SceneCreateRequest' }
      responses:
        '201':
          description: Scene created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneResponse' }
  /v1/scenes/{scene_id}:
    get:
      operationId: getScene
      tags: [scenes]
      summary: Get scene
      description: Get details of a specific scene by ID
      parameters:
        - in: path
          name: scene_id
          description: Scene identifier
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Scene details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneResponse' }
        '404':
          description: Scene not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      operationId: updateScene
      tags: [scenes]
      summary: Update scene
      description: Update an existing scene configuration
      parameters:
        - in: path
          name: scene_id
          description: Scene identifier
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SceneUpdateRequest' }
      responses:
        '200':
          description: Scene updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneResponse' }
        '404':
          description: Scene not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      operationId: deleteScene
      tags: [scenes]
      summary: Delete scene
      description: Delete a scene and its associated data
      parameters:
        - in: path
          name: scene_id
          description: Scene identifier
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Scene deleted
        '404':
          description: Scene not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/scenes/{scene_id}/execute:
    post:
      operationId: executeScene
      tags: [scenes]
      summary: Execute a scene
      description: Trigger execution of a scene with optional overrides
      parameters:
        - in: path
          name: scene_id
          description: Scene identifier
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SceneExecuteRequest' }
      responses:
        '202':
          description: Scene execution accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneExecuteResponse' }
  /v1/scenes/{scene_id}/executions:
    get:
      operationId: listSceneExecutions
      tags: [scenes]
      summary: List executions for scene
      description: Get execution history for a specific scene
      parameters:
        - in: path
          name: scene_id
          description: Scene identifier
          required: true
          schema: { type: string }
        - in: query
          name: limit
          description: Maximum number of executions to return
          schema: { type: integer }
        - in: query
          name: offset
          description: Number of executions to skip for pagination
          schema: { type: integer }
      responses:
        '200':
          description: List of scene executions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneExecutionsListResponse' }
  /v1/scenes/{scene_id}/stop:
    post:
      operationId: stopScenePlayback
      tags: [scenes]
      summary: Stop scene playback
      description: Stop playback for all devices in a scene
      parameters:
        - in: path
          name: scene_id
          description: Scene identifier
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Playback stopped on all devices
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScenePlaybackResponse' }
        '207':
          description: Partial success - some devices failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScenePlaybackResponse' }

  # =========================================================================
  # ROUTINES ENDPOINTS
  # =========================================================================
...
  /v1/executions:
    get:
      operationId: listExecutions
      tags: [executions]
      summary: List execution history
      description: Get paginated list of routine execution history
      parameters:
        - in: query
          name: limit
          description: Maximum number of executions to return
          schema: { type: integer }
        - in: query
          name: offset
          description: Number of executions to skip for pagination
          schema: { type: integer }
        - in: query
          name: routine_id
          description: Filter by routine identifier
          schema: { type: string }
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionHistoryResponse' }
...
components:
  schemas:
...
    Remediation:
      type: object
      description: Remediation guidance for error recovery
      required: [action]
      properties:
        action:
          type: string
          description: Machine-readable action code (e.g., 'BOOTSTRAP_SERVICE', 'RETRY_LATER')
        endpoint:
          type: string
          description: API endpoint to call to fix the issue
        user_action:
          type: string
          description: Human-readable instruction for the user
    ErrorResponse:
      type: object
      required: [request_id, error]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            remediation: { $ref: '#/components/schemas/Remediation' }
...
    ExecutionHistoryResponse:
      type: object
      required: [request_id, executions, pagination]
      properties:
        request_id: { type: string }
        executions:
          type: array
          items:
            type: object
            required:
              [
                id,
                routine_id,
                routine_name,
                timestamp,
                outcome,
                target_devices,
                content_played,
                failure_reason,
                failure_message,
                fallback_used
              ]
            properties:
              id: { type: string }
              routine_id: { type: string }
              routine_name: { type: string }
              timestamp: { type: string, format: date-time }
              outcome:
                type: string
                enum: [success, partial, failed, skipped, snoozed]
              target_devices:
                type: array
                items: { type: string }
              content_played:
                type: object
                additionalProperties: true
                nullable: true
              failure_reason:
                type: string
                nullable: true
              failure_message:
                type: string
                nullable: true
              fallback_used: { type: boolean }
        pagination:
          type: object
          required: [limit, offset, has_more]
          properties:
            limit: { type: integer }
            offset: { type: integer }
            has_more: { type: boolean }
...
    SceneResponse:
      type: object
      required: [request_id, scene]
      properties:
        request_id: { type: string }
        scene: { $ref: '#/components/schemas/Scene' }
    ScenesListResponse:
      type: object
      required: [request_id, scenes, pagination]
      properties:
        request_id: { type: string }
        scenes:
          type: array
          items: { $ref: '#/components/schemas/Scene' }
        pagination:
          type: object
          required: [total, limit, offset, has_more]
          properties:
            total: { type: integer }
            limit: { type: integer }
            offset: { type: integer }
            has_more: { type: boolean }
    SceneMember:
      type: object
      required: [device_id]
      properties:
        device_id: { type: string }
        room_name:
          type: string
          nullable: true
        target_volume:
          type: integer
          nullable: true
        mute:
          type: boolean
          nullable: true
    VolumeRamp:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        duration_ms:
          type: integer
          nullable: true
        curve:
          type: string
          enum: [linear, ease-in, ease-out]
          nullable: true
    Teardown:
      type: object
      properties:
        ungroup_after_ms:
          type: integer
          nullable: true
        restore_previous_state:
          type: boolean
          nullable: true
    Scene:
      type: object
      required:
        [
          scene_id,
          name,
          description,
          coordinator_preference,
          fallback_policy,
          members,
          volume_ramp,
          teardown,
          created_at,
          updated_at
        ]
      properties:
        scene_id: { type: string }
        name: { type: string }
        description:
          type: string
          nullable: true
        coordinator_preference:
          type: string
          enum: [ARC_FIRST]
        fallback_policy:
          type: string
          enum: [PLAYBASE_IF_ARC_TV_ACTIVE]
        members:
          type: array
          items: { $ref: '#/components/schemas/SceneMember' }
        volume_ramp:
          allOf:
            - $ref: '#/components/schemas/VolumeRamp'
          nullable: true
        teardown:
          allOf:
            - $ref: '#/components/schemas/Teardown'
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SceneCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        coordinator_preference:
          type: string
          enum: [ARC_FIRST]
        fallback_policy:
          type: string
          enum: [PLAYBASE_IF_ARC_TV_ACTIVE]
        members:
          type: array
          items: { $ref: '#/components/schemas/SceneMember' }
        volume_ramp: { $ref: '#/components/schemas/VolumeRamp' }
        teardown: { $ref: '#/components/schemas/Teardown' }
    SceneUpdateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        members:
          type: array
          items: { $ref: '#/components/schemas/SceneMember' }
        volume_ramp: { $ref: '#/components/schemas/VolumeRamp' }
        teardown: { $ref: '#/components/schemas/Teardown' }
    SceneExecution:
      type: object
      required:
        [
          scene_execution_id,
          scene_id,
          idempotency_key,
          coordinator_used_device_id,
          status,
          started_at,
          ended_at,
          steps,
          verification,
          error
        ]
      properties:
        scene_execution_id: { type: string }
        scene_id: { type: string }
        idempotency_key:
          type: string
          nullable: true
        coordinator_used_device_id:
          type: string
          nullable: true
        status:
          type: string
          enum: [STARTING, PLAYING_CONFIRMED, FAILED, ROLLED_BACK]
        started_at: { type: string, format: date-time }
        ended_at:
          type: string
          format: date-time
          nullable: true
        steps:
          type: array
          items: { $ref: '#/components/schemas/ExecutionStep' }
        verification:
          allOf:
            - $ref: '#/components/schemas/PlaybackVerification'
          nullable: true
        error:
          type: string
          nullable: true
    ExecutionStep:
      type: object
      required: [step, status]
      properties:
        step: { type: string }
        status: { type: string }
        started_at:
          type: string
          format: date-time
          nullable: true
        ended_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
    PlaybackVerification:
      type: object
      required: [playback_confirmed, checked_at]
      properties:
        playback_confirmed: { type: boolean }
        transport_state: { type: string }
        track_uri: { type: string }
        checked_at: { type: string, format: date-time }
        verification_unavailable: { type: boolean }
...
    SceneExecutionsListResponse:
      type: object
      required: [request_id, executions, pagination]
      properties:
        request_id: { type: string }
        executions:
          type: array
          items: { $ref: '#/components/schemas/SceneExecution' }
        pagination:
          type: object
          required: [total, limit, offset, has_more]
          properties:
            total: { type: integer }
            limit: { type: integer }
            offset: { type: integer }
            has_more: { type: boolean }
    ScenePlaybackResponse:
      type: object
      required: [request_id, result]
      properties:
        request_id: { type: string }
        result:
          type: object
          required: [scene_id, results, all_succeeded]
          properties:
            scene_id: { type: string }
            results:
              type: array
              items:
                type: object
                required: [device_id, success]
                properties:
                  device_id: { type: string }
                  success: { type: boolean }
                  error:
                    type: string
                    nullable: true
            all_succeeded: { type: boolean }
            started_at:
              type: string
              format: date-time
            stopped_at:
              type: string
              format: date-time
    SceneExecuteRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [MANUAL, SCHEDULED]
          description: Why the scene is being executed
        dry_run:
          type: boolean
          description: If true, validate but don't actually execute
        favorite_id:
          type: string
          deprecated: true
          description: |
            DEPRECATED - Use music_content instead.
            Legacy Sonos favorite ID (e.g., "FV:2/34") to play.
        content:
          $ref: '#/components/schemas/MusicContent'
          description: Music content to play (new format - preferred)
        queue_mode:
          $ref: '#/components/schemas/QueueMode'
          description: How to handle the playback queue
        group_behavior:
          $ref: '#/components/schemas/GroupBehavior'
          description: How to handle grouped speakers
        music_target:
          type: object
          deprecated: true
          description: |
            DEPRECATED - Use content instead.
            Legacy music target specification.
          additionalProperties: true
    SceneExecuteResponse:
      type: object
      required: [request_id, scene_execution_id]
      properties:
        request_id: { type: string }
        scene_execution_id: { type: string }
        content_played:
          $ref: '#/components/schemas/MusicContent'
          description: The music content that was played (if any)
        queue_mode_used:
          $ref: '#/components/schemas/QueueMode'
        group_behavior_used:
          $ref: '#/components/schemas/GroupBehavior'
...
    MusicService:
      type: string
      enum: [spotify, apple_music, amazon_music]
      description: Supported music streaming services
    MusicContentType:
      type: string
      enum: [playlist, album, station, track]
      description: Types of music content that can be played
    QueueMode:
      type: string
      enum: [REPLACE_AND_PLAY, PLAY_NEXT, ADD_TO_END, QUEUE_ONLY]
      default: REPLACE_AND_PLAY
      description: |
        How content is added to the Sonos queue:
        - REPLACE_AND_PLAY: Clear queue, add content, play immediately (default)
        - PLAY_NEXT: Insert after currently playing track
        - ADD_TO_END: Append to end of queue
        - QUEUE_ONLY: Add to queue without affecting playback state
    GroupBehavior:
      type: string
      enum: [AUTO_REDIRECT, UNGROUP_AND_PLAY, REQUIRE_COORDINATOR]
      default: AUTO_REDIRECT
      description: |
        How grouped speakers are handled:
        - AUTO_REDIRECT: Automatically route commands to coordinator (default)
        - UNGROUP_AND_PLAY: Ungroup the device first, then play on it alone
        - REQUIRE_COORDINATOR: Return error if device is not coordinator
    SonosFavoriteContent:
      type: object
      required: [type, favoriteId]
      properties:
        type:
          type: string
          enum: [sonos_favorite]
        favoriteId:
          type: string
          description: Sonos favorite ID (e.g., "FV:2/34")
    DirectContent:
      type: object
      required: [type, service, contentType, contentId, title]
      properties:
        type:
          type: string
          enum: [direct]
        service:
          $ref: '#/components/schemas/MusicService'
        contentType:
          $ref: '#/components/schemas/MusicContentType'
        contentId:
          type: string
          description: Service-specific content ID (e.g., Spotify playlist ID)
        title:
          type: string
          description: Display title for UI
        artworkUrl:
          type: string
          format: uri
          description: Optional artwork URL
    MusicContent:
      oneOf:
        - $ref: '#/components/schemas/SonosFavoriteContent'
        - $ref: '#/components/schemas/DirectContent'
      discriminator:
        propertyName: type
        mapping:
          sonos_favorite: '#/components/schemas/SonosFavoriteContent'
          direct: '#/components/schemas/DirectContent'
...